renamed ChargerTreads to Charger
