+++ src/main/java/frc/robot/RobotPreferences.java
@@ -52,7 +52,7 @@ public class RobotPreferences {
         "autoMaxAccelFeet", 1.0);
   }
 
-  public static final class prefChargerTreads {
+  public static final class prefCharger {
     public static final SN_DoublePreference chargerSpeed = new SN_DoublePreference("chargerSpeed", 0.25);
 
   }
